+++ src/datetime.rs
@@ -108,5 +108,9 @@ where
     }
 
 
+    //pub fn get_century()
+
+    //pub fn set_century()
+
 
 }